added base tests for tpu

+++ tests/base/deterministic_model.py
@@ -19,7 +19,7 @@ class DeterministicModel(LightningModule):
                 [4, 3, 5],
                 [10, 11, 13]
             ]).float()
-        self.l1 = torch.nn.Parameter(weights, requires_grad=True)
+        self.l1 = torch.nn.Parameter(weights, requires_grad=True).float()
 
     def forward(self, x):
         return self.l1.mm(x.t())
